requirements.txt
requests==2.31.0
schedule==1.2.0
plyer==2.1.0
python-dateutil==2.8.2
